Free Transform tool's handle should stop when it reaches the other end.
IKNINJA-1335 - Free Transform tool: When one of the handles reaches the other end, the tool should stop.
IKNINJA-1261 - [WebGL] [Shape] [Selection] Material shows in the middle of Canvas area if the shape is transformed toward negative coordinate. (This was happening because setting a canvas' width to 0 resets the canvas' width to 300.)

Signed-off-by: Nivesh Rajbhandari <mqg734@motorola.com>