Re-adding temp doc type
Adding a demo fix to include the document type, it is removed because comments are not supported currently, so we must added in manually. The YouTube component breaks, it does not support any doctype, hence, the component must have a bug filed against it, we MUST include doc types with our files.