IKNINJA-1671 - GIO: 3D rotate widget is offset from the object when rotated in negative space.
I fixed this temporarily by disabling the auto stage scroll calculations when we modify matrices. The correct fix will be to update any matrices and local2Global caches if we update the stage's padding (which I'm doing in a separate fix for a single draw cycle for the stage, but that fix is not ready yet).

Signed-off-by: Nivesh Rajbhandari <mqg734@motorola.com>