Support for browser escaped file paths (CSS)
This fixes an issue with files located in folder with names/paths that the browser escapes, hence, caused a bug with the path passed to the cloud simulator and files not to be found. This does not require testing to merge.
The fix is simple, addresses folder paths with spaces and such characters in which CSS files consumed by the HTML we open in design view.